+++ inlang/source-code/paraglide-examples/adapter-sveltekit/.svelte-kit/output/client/_app/immutable/chunks/index.562a46f0.js
@@ -1,0 +1,469 @@
+var __defProp = Object.defineProperty;
+var __defNormalProp = (obj, key, value) => key in obj ? __defProp(obj, key, { enumerable: true, configurable: true, writable: true, value }) : obj[key] = value;
+var __publicField = (obj, key, value) => {
+  __defNormalProp(obj, typeof key !== "symbol" ? key + "" : key, value);
+  return value;
+};
+import { r as run_all, n as noop, i as blank_object, j as flush, k as is_function, l as is_empty, m as add_render_callback, p as flush_render_callbacks, q as current_component, v as set_current_component, w as run, x as dirty_components, y as schedule_update } from "./scheduler.b51ce850.js";
+let is_hydrating = false;
+function start_hydrating() {
+  is_hydrating = true;
+}
+function end_hydrating() {
+  is_hydrating = false;
+}
+function upper_bound(low, high, key, value) {
+  while (low < high) {
+    const mid = low + (high - low >> 1);
+    if (key(mid) <= value) {
+      low = mid + 1;
+    } else {
+      high = mid;
+    }
+  }
+  return low;
+}
+function init_hydrate(target) {
+  if (target.hydrate_init)
+    return;
+  target.hydrate_init = true;
+  let children2 = (
+    /** @type {ArrayLike<NodeEx2>} */
+    target.childNodes
+  );
+  if (target.nodeName === "HEAD") {
+    const myChildren = [];
+    for (let i = 0; i < children2.length; i++) {
+      const node = children2[i];
+      if (node.claim_order !== void 0) {
+        myChildren.push(node);
+      }
+    }
+    children2 = myChildren;
+  }
+  const m = new Int32Array(children2.length + 1);
+  const p = new Int32Array(children2.length);
+  m[0] = -1;
+  let longest = 0;
+  for (let i = 0; i < children2.length; i++) {
+    const current = children2[i].claim_order;
+    const seqLen = (longest > 0 && children2[m[longest]].claim_order <= current ? longest + 1 : upper_bound(1, longest, (idx) => children2[m[idx]].claim_order, current)) - 1;
+    p[i] = m[seqLen] + 1;
+    const newLen = seqLen + 1;
+    m[newLen] = i;
+    longest = Math.max(newLen, longest);
+  }
+  const lis = [];
+  const toMove = [];
+  let last = children2.length - 1;
+  for (let cur = m[longest] + 1; cur != 0; cur = p[cur - 1]) {
+    lis.push(children2[cur - 1]);
+    for (; last >= cur; last--) {
+      toMove.push(children2[last]);
+    }
+    last--;
+  }
+  for (; last >= 0; last--) {
+    toMove.push(children2[last]);
+  }
+  lis.reverse();
+  toMove.sort((a, b) => a.claim_order - b.claim_order);
+  for (let i = 0, j = 0; i < toMove.length; i++) {
+    while (j < lis.length && toMove[i].claim_order >= lis[j].claim_order) {
+      j++;
+    }
+    const anchor = j < lis.length ? lis[j] : null;
+    target.insertBefore(toMove[i], anchor);
+  }
+}
+function append_hydration(target, node) {
+  if (is_hydrating) {
+    init_hydrate(target);
+    if (target.actual_end_child === void 0 || target.actual_end_child !== null && target.actual_end_child.parentNode !== target) {
+      target.actual_end_child = target.firstChild;
+    }
+    while (target.actual_end_child !== null && target.actual_end_child.claim_order === void 0) {
+      target.actual_end_child = target.actual_end_child.nextSibling;
+    }
+    if (node !== target.actual_end_child) {
+      if (node.claim_order !== void 0 || node.parentNode !== target) {
+        target.insertBefore(node, target.actual_end_child);
+      }
+    } else {
+      target.actual_end_child = node.nextSibling;
+    }
+  } else if (node.parentNode !== target || node.nextSibling !== null) {
+    target.appendChild(node);
+  }
+}
+function insert_hydration(target, node, anchor) {
+  if (is_hydrating && !anchor) {
+    append_hydration(target, node);
+  } else if (node.parentNode !== target || node.nextSibling != anchor) {
+    target.insertBefore(node, anchor || null);
+  }
+}
+function detach(node) {
+  if (node.parentNode) {
+    node.parentNode.removeChild(node);
+  }
+}
+function element(name) {
+  return document.createElement(name);
+}
+function text(data) {
+  return document.createTextNode(data);
+}
+function space() {
+  return text(" ");
+}
+function empty() {
+  return text("");
+}
+function listen(node, event, handler, options) {
+  node.addEventListener(event, handler, options);
+  return () => node.removeEventListener(event, handler, options);
+}
+function attr(node, attribute, value) {
+  if (value == null)
+    node.removeAttribute(attribute);
+  else if (node.getAttribute(attribute) !== value)
+    node.setAttribute(attribute, value);
+}
+function get_svelte_dataset(node) {
+  return node.dataset.svelteH;
+}
+function children(element2) {
+  return Array.from(element2.childNodes);
+}
+function init_claim_info(nodes) {
+  if (nodes.claim_info === void 0) {
+    nodes.claim_info = { last_index: 0, total_claimed: 0 };
+  }
+}
+function claim_node(nodes, predicate, processNode, createNode, dontUpdateLastIndex = false) {
+  init_claim_info(nodes);
+  const resultNode = (() => {
+    for (let i = nodes.claim_info.last_index; i < nodes.length; i++) {
+      const node = nodes[i];
+      if (predicate(node)) {
+        const replacement = processNode(node);
+        if (replacement === void 0) {
+          nodes.splice(i, 1);
+        } else {
+          nodes[i] = replacement;
+        }
+        if (!dontUpdateLastIndex) {
+          nodes.claim_info.last_index = i;
+        }
+        return node;
+      }
+    }
+    for (let i = nodes.claim_info.last_index - 1; i >= 0; i--) {
+      const node = nodes[i];
+      if (predicate(node)) {
+        const replacement = processNode(node);
+        if (replacement === void 0) {
+          nodes.splice(i, 1);
+        } else {
+          nodes[i] = replacement;
+        }
+        if (!dontUpdateLastIndex) {
+          nodes.claim_info.last_index = i;
+        } else if (replacement === void 0) {
+          nodes.claim_info.last_index--;
+        }
+        return node;
+      }
+    }
+    return createNode();
+  })();
+  resultNode.claim_order = nodes.claim_info.total_claimed;
+  nodes.claim_info.total_claimed += 1;
+  return resultNode;
+}
+function claim_element_base(nodes, name, attributes, create_element) {
+  return claim_node(
+    nodes,
+    /** @returns {node is Element | SVGElement} */
+    (node) => node.nodeName === name,
+    /** @param {Element} node */
+    (node) => {
+      const remove = [];
+      for (let j = 0; j < node.attributes.length; j++) {
+        const attribute = node.attributes[j];
+        if (!attributes[attribute.name]) {
+          remove.push(attribute.name);
+        }
+      }
+      remove.forEach((v) => node.removeAttribute(v));
+      return void 0;
+    },
+    () => create_element(name)
+  );
+}
+function claim_element(nodes, name, attributes) {
+  return claim_element_base(nodes, name, attributes, element);
+}
+function claim_text(nodes, data) {
+  return claim_node(
+    nodes,
+    /** @returns {node is Text} */
+    (node) => node.nodeType === 3,
+    /** @param {Text} node */
+    (node) => {
+      const dataStr = "" + data;
+      if (node.data.startsWith(dataStr)) {
+        if (node.data.length !== dataStr.length) {
+          return node.splitText(dataStr.length);
+        }
+      } else {
+        node.data = dataStr;
+      }
+    },
+    () => text(data),
+    true
+    // Text nodes should not update last index since it is likely not worth it to eliminate an increasing subsequence of actual elements
+  );
+}
+function claim_space(nodes) {
+  return claim_text(nodes, " ");
+}
+function set_data(text2, data) {
+  data = "" + data;
+  if (text2.data === data)
+    return;
+  text2.data = /** @type {string} */
+  data;
+}
+function set_style(node, key, value, important) {
+  if (value == null) {
+    node.style.removeProperty(key);
+  } else {
+    node.style.setProperty(key, value, important ? "important" : "");
+  }
+}
+function construct_svelte_component(component, props) {
+  return new component(props);
+}
+const outroing = /* @__PURE__ */ new Set();
+let outros;
+function group_outros() {
+  outros = {
+    r: 0,
+    c: [],
+    p: outros
+    // parent group
+  };
+}
+function check_outros() {
+  if (!outros.r) {
+    run_all(outros.c);
+  }
+  outros = outros.p;
+}
+function transition_in(block, local) {
+  if (block && block.i) {
+    outroing.delete(block);
+    block.i(local);
+  }
+}
+function transition_out(block, local, detach2, callback) {
+  if (block && block.o) {
+    if (outroing.has(block))
+      return;
+    outroing.add(block);
+    outros.c.push(() => {
+      outroing.delete(block);
+      if (callback) {
+        if (detach2)
+          block.d(1);
+        callback();
+      }
+    });
+    block.o(local);
+  } else if (callback) {
+    callback();
+  }
+}
+function create_component(block) {
+  block && block.c();
+}
+function claim_component(block, parent_nodes) {
+  block && block.l(parent_nodes);
+}
+function mount_component(component, target, anchor) {
+  const { fragment, after_update } = component.$$;
+  fragment && fragment.m(target, anchor);
+  add_render_callback(() => {
+    const new_on_destroy = component.$$.on_mount.map(run).filter(is_function);
+    if (component.$$.on_destroy) {
+      component.$$.on_destroy.push(...new_on_destroy);
+    } else {
+      run_all(new_on_destroy);
+    }
+    component.$$.on_mount = [];
+  });
+  after_update.forEach(add_render_callback);
+}
+function destroy_component(component, detaching) {
+  const $$ = component.$$;
+  if ($$.fragment !== null) {
+    flush_render_callbacks($$.after_update);
+    run_all($$.on_destroy);
+    $$.fragment && $$.fragment.d(detaching);
+    $$.on_destroy = $$.fragment = null;
+    $$.ctx = [];
+  }
+}
+function make_dirty(component, i) {
+  if (component.$$.dirty[0] === -1) {
+    dirty_components.push(component);
+    schedule_update();
+    component.$$.dirty.fill(0);
+  }
+  component.$$.dirty[i / 31 | 0] |= 1 << i % 31;
+}
+function init(component, options, instance, create_fragment, not_equal, props, append_styles, dirty = [-1]) {
+  const parent_component = current_component;
+  set_current_component(component);
+  const $$ = component.$$ = {
+    fragment: null,
+    ctx: [],
+    // state
+    props,
+    update: noop,
+    not_equal,
+    bound: blank_object(),
+    // lifecycle
+    on_mount: [],
+    on_destroy: [],
+    on_disconnect: [],
+    before_update: [],
+    after_update: [],
+    context: new Map(options.context || (parent_component ? parent_component.$$.context : [])),
+    // everything else
+    callbacks: blank_object(),
+    dirty,
+    skip_bound: false,
+    root: options.target || parent_component.$$.root
+  };
+  append_styles && append_styles($$.root);
+  let ready = false;
+  $$.ctx = instance ? instance(component, options.props || {}, (i, ret, ...rest) => {
+    const value = rest.length ? rest[0] : ret;
+    if ($$.ctx && not_equal($$.ctx[i], $$.ctx[i] = value)) {
+      if (!$$.skip_bound && $$.bound[i])
+        $$.bound[i](value);
+      if (ready)
+        make_dirty(component, i);
+    }
+    return ret;
+  }) : [];
+  $$.update();
+  ready = true;
+  run_all($$.before_update);
+  $$.fragment = create_fragment ? create_fragment($$.ctx) : false;
+  if (options.target) {
+    if (options.hydrate) {
+      start_hydrating();
+      const nodes = children(options.target);
+      $$.fragment && $$.fragment.l(nodes);
+      nodes.forEach(detach);
+    } else {
+      $$.fragment && $$.fragment.c();
+    }
+    if (options.intro)
+      transition_in(component.$$.fragment);
+    mount_component(component, options.target, options.anchor);
+    end_hydrating();
+    flush();
+  }
+  set_current_component(parent_component);
+}
+class SvelteComponent {
+  constructor() {
+    /**
+     * ### PRIVATE API
+     *
+     * Do not use, may change at any time
+     *
+     * @type {any}
+     */
+    __publicField(this, "$$");
+    /**
+     * ### PRIVATE API
+     *
+     * Do not use, may change at any time
+     *
+     * @type {any}
+     */
+    __publicField(this, "$$set");
+  }
+  /** @returns {void} */
+  $destroy() {
+    destroy_component(this, 1);
+    this.$destroy = noop;
+  }
+  /**
+   * @template {Extract<keyof Events, string>} K
+   * @param {K} type
+   * @param {((e: Events[K]) => void) | null | undefined} callback
+   * @returns {() => void}
+   */
+  $on(type, callback) {
+    if (!is_function(callback)) {
+      return noop;
+    }
+    const callbacks = this.$$.callbacks[type] || (this.$$.callbacks[type] = []);
+    callbacks.push(callback);
+    return () => {
+      const index = callbacks.indexOf(callback);
+      if (index !== -1)
+        callbacks.splice(index, 1);
+    };
+  }
+  /**
+   * @param {Partial<Props>} props
+   * @returns {void}
+   */
+  $set(props) {
+    if (this.$$set && !is_empty(props)) {
+      this.$$.skip_bound = true;
+      this.$$set(props);
+      this.$$.skip_bound = false;
+    }
+  }
+}
+const PUBLIC_VERSION = "4";
+if (typeof window !== "undefined")
+  (window.__svelte || (window.__svelte = { v: /* @__PURE__ */ new Set() })).v.add(PUBLIC_VERSION);
+export {
+  SvelteComponent as S,
+  insert_hydration as a,
+  check_outros as b,
+  claim_space as c,
+  transition_in as d,
+  empty as e,
+  detach as f,
+  element as g,
+  claim_element as h,
+  init as i,
+  children as j,
+  attr as k,
+  set_style as l,
+  text as m,
+  claim_text as n,
+  set_data as o,
+  group_outros as p,
+  construct_svelte_component as q,
+  create_component as r,
+  space as s,
+  transition_out as t,
+  claim_component as u,
+  mount_component as v,
+  destroy_component as w,
+  append_hydration as x,
+  get_svelte_dataset as y,
+  listen as z
+};