+++ inlang/source-code/paraglide/paraglide-js-adapter-webpack/src/index.ts
@@ -1,0 +1,2 @@
+import { paraglide as unpluginParaglide } from "@inlang/paraglide-js-adapter-unplugin"
+export default unpluginParaglide.webpack