+++ inlang/source-code/paraglide/paraglide-js/src/compiler/compile.test.ts
@@ -22,11 +22,14 @@ test("the files should include a prettierignore file", async () => {
 	expect(output).toHaveProperty(".prettierignore")
 	expect(output[".prettierignore"]).toContain("*")
 })
-// ignore eslint stuff
-test("files should include an eslint ignore", async () => {
-	expect(output).toHaveProperty(".eslintignore")
-	expect(output[".eslintignore"]).toContain("*")
-})
+
+// All JS files must be eslint ignored
+test("the files should include an eslint ignore comment", async () => {
+	for (const [filePath, content] of Object.entries(output)) {
+		if (!filePath.endsWith(".js")) continue;
+		expect(content).toContain("/* eslint-disable */")
+	}
+});
 
 test("imports in the messages.js index file should use underscores instead of hyphens to avoid invalid JS imports", async () => {
 	expect(output["messages.js"]).toContain("import * as en_US")
@@ -221,6 +224,30 @@ describe("e2e", async () => {
 			runtime.onSetLanguageTag(() => {})
 		}).toThrow()
 	})
+
+	test("should return the correct message if a languageTag is set in the message options", async () => {
+		const { m, runtime } = await import(
+			`data:application/javascript;base64,${Buffer.from(
+				compiledBundle.output[0].code,
+				"utf8"
+			).toString("base64")}`
+		)
+
+		// set the language tag to de to make sure that the message options override the runtime language tag
+		runtime.setLanguageTag("de")
+		expect(m.onlyText()).toBe("Eine einfache Nachricht.")
+		expect(m.onlyText(undefined, { languageTag: "en" })).toBe("A simple message.")
+		expect(m.multipleParams({ name: "Samuel", count: 5 }, { languageTag: "en" })).toBe(
+			"Hello Samuel! You have 5 messages."
+		)
+
+		runtime.setLanguageTag("en")
+		expect(m.onlyText({}, { languageTag: "de" })).toBe("Eine einfache Nachricht.")
+		expect(m.oneParam({ name: "Samuel" }, { languageTag: "de" })).toBe("Guten Morgen Samuel!")
+		expect(m.multipleParams({ name: "Samuel", count: 5 }, { languageTag: "de" })).toBe(
+			"Hallo Samuel! Du hast 5 Nachrichten."
+		)
+	})
 })
 
 describe("tree-shaking", () => {
@@ -353,9 +380,8 @@ test("typesafety", async () => {
     // languageTag should return type should be a union of language tags, not a generic string
     runtime.languageTag() satisfies "de" | "en" | "en-US"
 
-		// setting the language tag as a getter function should be possible
-
-		runtime.setLanguageTag(() => "en")
+	// setting the language tag as a getter function should be possible
+	runtime.setLanguageTag(() => "en")
 
     // --------- MESSAGES ---------
 
@@ -370,6 +396,18 @@ test("typesafety", async () => {
 
     // a message without params shouldn't require params
     m.onlyText() satisfies string
+
+
+	// --------- MESSAGE OPTIONS ---------
+	// the languageTag option should be optional
+	m.onlyText({}, {}) satisfies string
+
+	// the languageTag option should be allowed
+	m.onlyText({}, { languageTag: "en" }) satisfies string
+
+	// the languageTag option must be a valid language tag
+	// @ts-expect-error - invalid language tag
+	m.onlyText({}, { languageTag: "---" })
   `
 	)